fixed annuity calculation
the calculation of simple interest and compound interest rate values were based on present values. this has coused the distortion in the bar chart representation. the simple interest rate seemed to diminish. although form a present value standpoint, this was not wrong entirely, this would confuse the user. Now the annuity stays in value from a future value standpoint, so it is throughout the periodes even (cummulative). Even are also the values of simple interest rates applied on the principal.